requirements.txt
rich
pyfiglet
PyYAML